backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.6,<5.0
psycopg[binary]>=3.2,<4.0
python-decouple>=3.8,<4.0
drf-spectacular>=0.28,<1.0
celery>=5.4,<6.0
redis>=5.2,<6.0
django-filter>=24.3,<25.0
Pillow>=11.1,<12.0
